Separation of the activities view and the picture view
