+++ app/build.gradle
@@ -106,7 +106,6 @@ dependencies {
 
     implementation 'com.airbnb.android:lottie:3.7.0'
     implementation 'com.github.MikeOrtiz:TouchImageView:3.1.1'
-    implementation 'com.github.rubensousa:gravitysnaphelper:2.2.1'
     implementation 'com.tbuonomo.andrui:viewpagerdotsindicator:4.1.2'
     implementation 'com.wdullaer:materialdatetimepicker:4.2.3'
