+++ app/src/main/res/layout/fragment_home.xml
@@ -79,6 +79,7 @@
                         android:layout_height="wrap_content"
                         android:layout_marginHorizontal="@dimen/marginStandard"
                         android:layout_marginTop="@dimen/marginStandardMedium"
+                        android:layout_marginBottom="@dimen/marginStandardSmall"
                         android:visibility="gone"
                         tools:visibility="visible" />
 
@@ -88,7 +89,7 @@
                         android:layout_width="match_parent"
                         android:layout_height="50dp"
                         android:layout_marginHorizontal="@dimen/marginStandard"
-                        android:layout_marginTop="@dimen/marginStandard"
+                        android:layout_marginTop="@dimen/marginStandardMedium"
                         android:layout_marginBottom="@dimen/marginStandardMedium" />
 
                     <include
@@ -110,31 +111,44 @@
                         android:visibility="gone"
                         tools:visibility="visible" />
 
-                    <TextView
-                        android:id="@+id/lastFilesTitle"
-                        style="@style/H2"
+                    <com.google.android.material.button.MaterialButtonToggleGroup
+                        android:id="@+id/tabsHomeGroup"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
                         android:layout_marginHorizontal="@dimen/marginStandard"
-                        android:layout_marginTop="@dimen/marginStandardSmall"
-                        android:text="@string/homeLastFilesTitle" />
-
-                    <com.github.rubensousa.gravitysnaphelper.GravitySnapRecyclerView
-                        android:id="@+id/lastFilesRecyclerView"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:clipToPadding="false"
-                        android:nestedScrollingEnabled="false"
                         android:orientation="horizontal"
-                        android:overScrollMode="never"
-                        android:paddingVertical="@dimen/marginStandardSmall"
-                        android:paddingStart="@dimen/marginStandard"
-                        android:paddingEnd="100dp"
-                        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                        app:snapGravity="start"
-                        app:snapToPadding="true"
-                        tools:ignore="RtlSymmetry"
-                        tools:listitem="@layout/cardview_file_grid" />
+                        app:checkedButton="@+id/homeActivitiesButton"
+                        app:layout_constraintEnd_toEndOf="parent"
+                        app:layout_constraintStart_toStartOf="parent"
+                        app:layout_constraintTop_toBottomOf="@id/searchViewCard"
+                        app:selectionRequired="true"
+                        app:singleSelection="true">
+
+                        <com.google.android.material.button.MaterialButton
+                            android:id="@+id/homeActivitiesButton"
+                            style="@style/ButtonToggleGroupAlternative"
+                            android:layout_width="0dp"
+                            android:layout_height="wrap_content"
+                            android:layout_weight="1"
+                            android:text="@string/fileDetailsActivitiesTitle" />
+
+                        <com.google.android.material.button.MaterialButton
+                            android:id="@+id/homeOfflineButton"
+                            style="@style/ButtonToggleGroupAlternative"
+                            android:layout_width="0dp"
+                            android:layout_height="wrap_content"
+                            android:layout_weight="1"
+                            android:text="@string/offlineFileTitle" />
+
+                        <com.google.android.material.button.MaterialButton
+                            android:id="@+id/homePicturesButton"
+                            style="@style/ButtonToggleGroupAlternative"
+                            android:layout_width="0dp"
+                            android:layout_height="wrap_content"
+                            android:layout_weight="1"
+                            android:text="@string/allPictures" />
+
+                    </com.google.android.material.button.MaterialButtonToggleGroup>
 
                 </LinearLayout>
 
@@ -142,42 +156,12 @@
 
         </com.google.android.material.appbar.AppBarLayout>
 
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/lastElementsLayout"
+        <androidx.viewpager2.widget.ViewPager2
+            android:id="@+id/homeViewPager"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:layout_marginTop="@dimen/marginStandardMedium"
-            android:orientation="vertical"
-            app:layout_behavior="@string/appbar_scrolling_view_behavior">
-
-            <TextView
-                android:id="@+id/lastElementsTitle"
-                style="@style/H2"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="@dimen/marginStandard"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintHorizontal_bias="0.0"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                tools:text="@string/homeMyLastPictures" />
-
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/lastElementsRecyclerView"
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:clipToPadding="false"
-                android:overScrollMode="never"
-                android:paddingHorizontal="20dp"
-                android:paddingTop="@dimen/marginStandardSmall"
-                android:paddingBottom="@dimen/recyclerViewPaddingBottom"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/lastElementsTitle"
-                tools:itemCount="5"
-                tools:listitem="@layout/cardview_home_file_activity" />
-        </androidx.constraintlayout.widget.ConstraintLayout>
+            app:layout_behavior="@string/appbar_scrolling_view_behavior" />
 
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
